slug,timestampable and data validation

+++ app/src/Controller/CategoryController.php
@@ -182,7 +182,15 @@ class CategoryController extends AbstractController
     )]
     public function delete(Request $request, Category $category): Response
     {
-        $form = $this->createForm(FormType::class, $category, [
+        if(!$this->categoryService->canBeDeleted($category)) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.category_contains_books')
+            );
+            return $this->redirectToRoute('book_index');
+        }
+
+            $form = $this->createForm(FormType::class, $category, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
         ]);
